Shorten button ids and attached functionality from inside <script> tag.
This is in the way of cleanup.  Changed button ids from, e.g. 'button-punch' to
'punch'.  I mean, we already know it's a button -- that's its element type.

Moving the functionality involved attaching the click function from inside the
page's <script> tag, instead of using the 'onclick' attribute of the element.
This seems like a better division of code, keeping markup and behavior
separate.